test: Add TabularDataset test for processing image data and generating dataset; fixtures changed
