+++ tests/datasets/test_datasets_tabular.py
@@ -3,9 +3,8 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from siapy.datasets import TabularDataset
 from siapy.datasets.tabular import DatasetDataFrame, TabularDataEntity
-from siapy.entities import Pixels, Shape, SpectralImageSet
+from siapy.entities import Pixels, SpectralImageSet
 
 
 @pytest.fixture(scope="module")
@@ -71,78 +70,46 @@ def test_pixels(sample_dataframe):
     )
 
 
-@pytest.fixture(scope="module")
-def spectral_images_set(spectral_images):
-    pixels_input = [(10, 15), (60, 66)]
-    pixels = Pixels.from_iterable(pixels_input)
-    rectangle = Shape.from_shape_type(shape_type="rectangle", pixels=pixels)
-
-    spectral_images.vnir.geometric_shapes.append(rectangle)
-    spectral_images.swir.geometric_shapes.append(rectangle)
-
-    images = [
-        spectral_images.vnir,
-        spectral_images.swir,
-        spectral_images.vnir,
-    ]
-
-    return SpectralImageSet(images)
-
-
-def test_tabular(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
-    dataset.generate_dataset()
-
-
-def test_tabular_len(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
+def test_tabular_len(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     assert len(dataset) == 3
 
 
-def test_tabular_str(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
+def test_tabular_str(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     expected_str = "<TabularDataset object with 3 data entities>"
     assert str(dataset) == expected_str
 
 
-def test_tabular_iter(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
+def test_tabular_iter(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     for entity in dataset:
         assert isinstance(entity, TabularDataEntity)
 
 
-def test_tabular_getitem(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
+def test_tabular_getitem(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     first_entity = dataset[0]
     assert isinstance(first_entity, TabularDataEntity)
 
 
-def test_tabular_image_set(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
+def test_tabular_image_set(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     assert isinstance(dataset.image_set, SpectralImageSet)
 
 
-def test_tabular_data_entities(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
+def test_tabular_data_entities(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     data_entities = dataset.data_entities
     assert all(isinstance(entity, TabularDataEntity) for entity in data_entities)
 
 
-def test_tabular_process_image_data(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
+def test_tabular_process_image_data(spectral_tabular_dataset):
+    dataset = spectral_tabular_dataset.dataset
     assert len(dataset.data_entities) > 0
 
 
-def test_tabular_generate_dataset(spectral_images_set):
-    dataset = TabularDataset(spectral_images_set)
-    dataset.process_image_data()
-    df = dataset.generate_dataset()
+def test_tabular_generate_dataset(spectral_tabular_dataset):
+    df = spectral_tabular_dataset.dataset_df
     assert isinstance(df, DatasetDataFrame)
     assert not df.empty
